+++ src/test/module.ts
@@ -1,0 +1,6 @@
+import { Module } from '@nestjs/common';
+import { ConsoleModule } from '../module';
+@Module({
+    imports: [ConsoleModule]
+})
+export class ConsoleModuleTest {}