+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]-[1.0.0-alpha.1] - 2019-04-08
+## [1.0.0]
+
+### Changed
+
+-   The bootstrap method has been moved and is now static in the new `BootstrapConsole` class, you can extend it if you need to change the default boot behavior.
+-   New options for bootstrap
+
+## [1.0.0-alpha.1] - 2019-04-08
 
 ### Added
 
@@ -40,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Update the dependencies to support nestjs ^6.0.0
 
-[unreleased]: https://github.com/Pop-Code/nestjs-console/compare/v1.0.0-alpha.1...HEAD
+[unreleased]: https://github.com/Pop-Code/nestjs-console/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/Pop-Code/nestjs-console/compare/v1.0.0...v1.0.0-alpha.1
 [1.0.0-alpha.1]: https://github.com/Pop-Code/nestjs-console/compare/v0.2.1...v1.0.0-alpha.1
 [0.2.2]: https://github.com/Pop-Code/nestjs-console/compare/v0.2.1...v1.0.0-alpha.1
 [0.2.1]: https://github.com/Pop-Code/nestjs-console/compare/v0.2.0...v0.2.1